Use new custom metrics collector
This uses the new custom KFP V1 metrics collector that can directly extract
metrics from Kubeflow Pipeline metrics.

With this collector, to measure metrics of a Kubeflow pipeline
only requires to a) add the label which step is the `model-training`
b) diseable caching for this step
c) configure the katib metrics collector.

Also all the information is added now, such that the Katib pipeline
can be run via the KatibClient.

Addresses:
- https://github.com/kubeflow/katib/issues/1914
- https://github.com/kubeflow/katib/issues/2019

k